+++ src/Cloud.Framework.Domain.Abstractions/Interfaces/IEvent.cs
@@ -1,3 +1,5 @@
+using System;
+
 namespace Cloud.Framework.Domain.Abstractions.Interfaces
 {
     /// <summary>
@@ -6,6 +8,5 @@ namespace Cloud.Framework.Domain.Abstractions.Interfaces
     /// <remarks>Has no methods/properties because it's just used to tag an object as an event to avoid external dependencies.</remarks>
     public interface IEvent
     {
-        
     }
 }