+++ src/Cloud.Framework.Domain.Abstractions/ChangeLog.md
@@ -12,3 +12,4 @@ All notable changes to the `Cloud.Framework.Domain.Abstractions` codebase.
   and `net5.0` as optional frameworks.
 - Moved most of the common project attributes to [common.props](../../common.props) file
   and imported it.
+- Updated [AggregateChild](./Base/AggregateChild.cs) `Parent` property to `protected set` from `private set`.